Correlates of War list of independent states

A list of independent states and microstates from 1816 on by the Correlates of War project.

cowstates

Format

Data frame

ccode

Gleditsch and Ward country code.

cowc

ISO 3 character country code.

country_name

Long form country name

start

Country start of independence.

end

Country end of independence.

microstate

Logical flag for whether state is a microstates with less than 250,000 population.

Source

Correlates of War Project. 2011. "State System Membership List, v2011." Online, https://correlatesofwar.org

Examples

data(cowstates)
head(cowstates)
#>   cowcode cowc             country_name      start        end microstate
#> 1       2  USA United States of America 1816-01-01 9999-12-31      FALSE
#> 2      20  CAN                   Canada 1920-01-10 9999-12-31      FALSE
#> 3      31  BHM                  Bahamas 1973-07-10 9999-12-31      FALSE
#> 4      40  CUB                     Cuba 1902-05-20 1906-09-25      FALSE
#> 5      40  CUB                     Cuba 1909-01-23 9999-12-31      FALSE
#> 6      41  HAI                    Haiti 1859-01-01 1915-07-28      FALSE