Gleditsch and Ward list of independent states

A list of independent states and microstates from 1816 on by Gleditsch and Ward

gwstates

Format

Data frame

gwcode

Gleditsch and Ward country code.

iso3c

ISO 3 character country code.

country_name

Long form country name

start

Country start of independence.

end

Country end of independence.

microstate

Logical flag for whether state is a microstates with less than 250,000 population.

Source

Gleditsch, Kristian S. and Michael D. Ward. 1999. ``Interstate System Membership: A Revised List of the Independent States since 1816." International Interactions 25.

Examples

data(gwstates)
head(gwstates)

#>   gwcode iso3c             country_name      start        end microstate
#> 1      2   USA United States of America 1816-01-01 2017-12-31      FALSE
#> 2     20   CAN                   Canada 1867-07-01 2017-12-31      FALSE
#> 3     31   BHM                  Bahamas 1973-07-10 2017-12-31      FALSE
#> 4     40   CUB                     Cuba 1902-05-20 2017-12-31      FALSE
#> 5     41   HAI                    Haiti 1816-01-01 1915-07-04      FALSE
#> 6     41   HAI                    Haiti 1934-08-15 2017-12-31      FALSE