Visualize missing status for state panel

Visualize missing status for state panel

plot_missing(x, data, space, time, time_unit, statelist = c("none", "GW",
  "COW"))

Arguments

x
Variable names(s), e.g. "x" or c("x1", "x2"). To check all variables in a data frame use something like.

data
State panel data frame

space
Name of state identifier

time
Name of time identifier

time_unit
Temporal resolution character string, e.g. "year" or "month". See details in seq.Date.

statelist
Check not only missing values, but presence or absence of observations against a list of independent states?

Examples

# To check all variables:
# plot_missing(setdiff(colnames(df), "space", "time"), ...)