Visualize missing and non-proper cases for state panel data

Plot missing values by country and date, and additionally identify country-date cases that do or do not match an independent state list.

plot_missing(data, x, space, time, time_unit, statelist = c("none", "GW",
  "COW"), skip_labels = 5)

missing_info(data, x, space, time, time_unit, statelist = "none")

Arguments

data
State panel data frame

x
Variable names(s), e.g. "x" or c("x1", "x2").

space
Name of variable identifying state country codes.

time
Name of time identifier, the corresponding variables needs to be Date class.

time_unit
Temporal resolution character string, e.g. "year" or "month". See details in seq.Date.

statelist
Check not only missing values, but presence or absence of observations against a list of independent states? "none" or "GW" or "COW".

skip_labels
Only plot the label for every n-th country on the y-axis to avoid overplotting.

Value

plot_missing returns a ggplot2 object.

missing_info returns a data frame with components:

[space]

Space identifier, with name equal to the "space" argument, e.g. "ccode".

[time]

Time identifier, with name equal to the "time" argument, e.g. "date".

independent

A logical vector, is the statelist argument is none, NA.

missing_value

A logical vector indicating if that record has missing values

status

The label used for plotting, combining the independence and missing value information for a case as appropriate.

Details

missing_info provides the information that is plotted with plot_missing. The latter returns a ggplot, and thus can be chained with other ggplot functions as usual.

Examples

# Create an example data frame with missing values
cy <- state_panel(as.Date("1980-06-30"), as.Date("2015-06-30"), by = "year",
useGW = TRUE)
cy$myvar <- rnorm(nrow(cy))
cy$myvar[sample(1:nrow(cy), nrow(cy)*.1, replace = FALSE)] <- NA
str(cy)

#> 'data.frame':	6654 obs. of  3 variables:
#>  $ gwcode: int  2 2 2 2 2 2 2 2 2 2 ...
#>  $ date  : Date, format: "1980-06-30" "1981-06-30" ...
#>  $ myvar : num  -1.40004 0.25532 -2.43726 -0.00557 0.62155 ...

# Visualize missing values:
plot_missing("myvar", cy, "gwcode", "date", "year", "none")

#> Error in `[.data.frame`(df, , space): undefined columns selected

# missing_info() generates the data underlying plot_missing():
head(missing_info("myvar", cy, "gwcode", "date", "year", "none"))

#> Error in `[.data.frame`(df, , space): undefined columns selected

# if we specify a statelist to check against, 'independent' will have values
# now:
head(missing_info("myvar", cy, "gwcode", "date", "year", "GW"))

#> Error in `[.data.frame`(df, , space): undefined columns selected

# Check data also against G&W list of independent states
head(missing_info("myvar", cy, "gwcode", "date", "year", "GW"))

#> Error in `[.data.frame`(df, , space): undefined columns selected

plot_missing("myvar", cy, "gwcode", "date", "year", "GW")

#> Error in `[.data.frame`(df, , space): undefined columns selected

# To check all variables:
# plot_missing(setdiff(colnames(df), "space", "time"), ...)

# Live example with Polity data
data("polity")
head(polity)

#>   ccode year polity
#> 1   700 1800     -6
#> 2   700 1801     -6
#> 3   700 1802     -6
#> 4   700 1803     -6
#> 5   700 1804     -6
#> 6   700 1805     -6

polity$date <- as.Date(paste0(polity$year, "-12-31"))
plot_missing("polity", polity, "ccode", "date", "year", "COW")

#> Error in `[.data.frame`(df, , space): undefined columns selected

# COW starts in 1816; Polity has excess data for several non-independent
# states after that date, and is missing coverage for several countries.

# The date option is relevant for years in which states gain or lose
# independence, so this will be slighlty different:
polity$date <- as.Date(paste0(polity$year, "-01-01"))
plot_missing("polity", polity, "ccode", "date", "year", "COW")

#> Error in `[.data.frame`(df, , space): undefined columns selected

# plot_missing returns a ggplot2 object, so you can do anything you want
plot_missing("polity", polity, "ccode", "date", "year", "COW") +
  ggplot2::coord_flip()

#> Error in `[.data.frame`(df, , space): undefined columns selected