Create state panel data

Create panel data consisting of independent states in the international system.

state_panel(start, end, by = "year", partial = "exact", useGW = TRUE)

Arguments

start
Beginning date for data

end
End date for data

by
Temporal resolution, "year", "month", or "day".

partial
Option for how to handle edge cases where a state is independent for only part of a time period (year, month, etc.). Options include "exact", and "any". See details.

useGW
Use Gleditsch & Ward statelist or Correlates of War state system membership list.

Details

The partial option determines how to handle instances where a country gains or loses independence during a time period specified in the by option:

"exact": the exact date in start is used for filtering

"any": a state-period is included if the state was independent at any point in that period.

Examples

gwlist  <- state_panel("1991-01-01", "2015-01-01", by = "year", useGW = TRUE)
head(gwlist, 3)

#>   gwcode       date
#> 1      2 1991-01-01
#> 2      2 1992-01-01
#> 3      2 1993-01-01

cowlist <- state_panel("1991-01-01", "2015-01-01", by = "year", useGW = FALSE)
head(cowlist, 3)

#>   cowcode       date
#> 1       2 1991-01-01
#> 2       2 1992-01-01
#> 3       2 1993-01-01

# Partials
# Focus on South Sudan--is there a record for 2011, first year of indendence?
data(gwstates)
dplyr::filter(gwstates, gwcode==626)

#>   gwcode iso3c country_name      start        end microstate
#> 1    626   SSD  South Sudan 2011-07-09 2016-12-31      FALSE

# No 2011 because SSD was not indpendent on January 1st 2011
x <- state_panel("2011-01-01", "2013-01-01", by = "year", partial = "exact")
dplyr::filter(x, gwcode==626)

#>   gwcode       date
#> 1    626 2012-01-01
#> 2    626 2013-01-01

# Includes 2011 because 12-31 date is used for filtering
x <- state_panel("2011-12-31", "2013-12-31", by = "year", partial = "exact")
dplyr::filter(x, gwcode==626)

#>   gwcode       date
#> 1    626 2011-12-31
#> 2    626 2012-12-31
#> 3    626 2013-12-31

# Includes 2011 because partial = "any"
x <- state_panel("2011-01-01", "2013-01-01", by = "year", partial = "any")
dplyr::filter(x, gwcode==626)

#>   gwcode       date
#> 1    626 2011-01-01
#> 2    626 2012-01-01
#> 3    626 2013-01-01